// src/app/core/models/statistique/stats-globales.ts
export interface DashboardStats {
  nombreEleves?: number;
  nombreEnseignant?: number;
  nombreClasse?: number;
  nombreEleveInscrits?: number;
  nombreEleveNonInscrits?: number;
  montantTotalEnAttante?: number;
  montantTotalEnRetard?: number;
  montantTotalImpaye?: number;
  nombreTotalFactureImpaye?: number;

}